extract-sms-email-data/requirements.txt
aiohttp>=3.8.0
tqdm>=4.64.0
python-dotenv>=0.19.0
pymongo>=4.0.0
psutil>=5.8.0
statistics
asyncio
